rename config.ts to lyraConfig.ts
