+++ webapp/src/utils/adapters/MessageAdapterFactory.ts
@@ -1,6 +1,6 @@
 import TsMessageAdapter from './TsMessageAdapter';
 import YamlMessageAdapter from './YamlMessageAdapter';
-import { LyraProjectConfig, MessageKind } from '../config';
+import { LyraProjectConfig, MessageKind } from '../lyraConfig';
 
 export default class MessageAdapterFactory {
   static createAdapter(config: LyraProjectConfig) {
